+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Minor changes on README. CHANGELOG also added. No functional changes
   MINOR `terminal_uuid` included in klap calls and `TapoDevice` handler includes possibility to call with a random uuid.
   BUG   Initial sporadic `negotiation errors`solved. Now 99.99% of initial message results in proper protocol when `AUTO`. Also when protocol is manually selected.
 
+- [tapo_actions.js](https://github.com/mbserran/node-red-contrib-tapo-new-api/blob/master/src/nodes/tapo_actions.ts) 
+  MAJOR Two more commands included: `components` to get device components in device object and `command` to send a customized request.
+
 - [tapo_actions.json](https://github.com/mbserran/node-red-contrib-tapo-new-api/blob/master/src/nodes/locales/es-ES/tapo_actions.json)
   New `es-ES` language support included